+++ tests/testthat/test-tree_rearrange.R
@@ -113,12 +113,46 @@ test_that("RootOnNode() works", {
 
 })
 
+
+test_that("root_on_node() works", {
+  tree <- Preorder(BalancedTree(15))
+  edge <- tree$edge
+  TipTest <- function (i) {
+    tr.rooted <- root_on_node(tree, i)
+    expect_equal(SortTree(root(tree, i, resolve.root = TRUE)),
+                 SortTree(tr.rooted))
+  }
+  StaticTest <- function (i) expect_equal(tree, root_on_node(tree, i))
+  NodeTest <- function (i) {
+    tr.rooted <- root_on_node(tree, i)
+    expect_equal(SortTree(root(tree, node = i, resolve.root = TRUE)),
+                 SortTree(tr.rooted))
+  }
+  expect_error(root_on_node(edge, 0))
+  for (i in 1:15) TipTest(i)
+  StaticTest(16)
+  StaticTest(17)
+  for (i in 18:23) NodeTest(i)
+  StaticTest(24)
+  for (i in 24:29) NodeTest(i)
+  expect_error(root_on_node(edge, 30))
+
+  tree <- Preorder(root(BalancedTree(15), 't1', resolve.root = TRUE))
+  edge <- tree$edge
+  expect_error(root_on_node(edge, 0))
+  for (i in 1:15) TipTest(i)
+  StaticTest(16)
+  StaticTest(17)
+  for (i in 18:29) NodeTest(i)
+  expect_error(root_on_node(edge, 30))
+})
+
 test_that("RootOnNode() supports lists of trees", {
   rootOn <- 8L
-  expect_equal(structure(list(RootOnNode(as.phylo(1, 5), rootOn),
+  expect_equivalent(structure(list(RootOnNode(as.phylo(1, 5), rootOn),
                               RootOnNode(as.phylo(2, 5), rootOn)),
                          class = 'multiPhylo'),
-               RootOnNode(as.phylo(1:2, 5), rootOn))
+                    RootOnNode(as.phylo(1:2, 5), rootOn))
 })
 
 test_that("RootTree() supports star trees", {
@@ -148,10 +182,10 @@ test_that("RootTree() works", {
                RootTree(bal8, c('t5', 't6')))
   expect_equal(RootTree(bal8, c('t1', 't2')), RootTree(bal8, c('t4', 't5')))
 
-  expect_equal(structure(list(RootTree(as.phylo(1, 5), 't5'),
-                              RootTree(as.phylo(2, 5), 't5')),
-                         class = 'multiPhylo'),
-               RootTree(as.phylo(1:2, 5), 't5'))
+  expect_equivalent(structure(list(RootTree(as.phylo(1, 5), 't5'),
+                                   RootTree(as.phylo(2, 5), 't5')),
+                              class = 'multiPhylo'),
+                    RootTree(as.phylo(1:2, 5), 't5'))
 
 })
 
@@ -173,7 +207,7 @@ test_that("UnrootTree() works", {
 
   expList <- list(UnrootTree(as.phylo(1, 5)), UnrootTree(as.phylo(2, 5)))
   expect_equal(expList, UnrootTree(list(as.phylo(1, 5), as.phylo(2, 5))))
-  expect_equal(structure(expList, class = 'multiPhylo'),
+  expect_equivalent(structure(expList, class = 'multiPhylo'),
                     UnrootTree(as.phylo(1:2, 5)))
 })